refactor(feat/trans_aligned): 重构代码，更符合工程化
